Made some minor improvements to the custom slider class.
Made it so when a slider is ran for the first time, it will now focus itself to the first available item in the menu to adjust. You can now press enter on a slider to set sertain values to it's defaults. This especially goes for volume and pitches of platforms, walls, musics, and sources.
Added the ability to press the page up and down keys to increase values by 10 in the slider class.