Entities can now be trapped in walls.
Introduced wall detection for entities such as animals and enemies. This means that if there's a wall anywhere on the map, the entitie won't be allowed to go past it what so ever. From now on, the only way they'll be able to pass through it is if the wall is removed from the map, or if a passage that contains 2 walls on both endpoints is opened. Destroyable walls are temporarily unaffected untill the next version.