Added and modified some slash commands found in the help file.
Added the ability to give yourself items with the slash give command.
Added the ability to collect all of the items found on the map with the slash collectall command.
Added the ability to move entities and other objects on the map with the slash go command, not just yourself.