Added some more bone related tweaks, along with other bug fixes and enhancements.
It is now possible to lose a life and or die from breaking too many bones in the game.
Made it so all of your bones are reset to being fully functional upon death or character reload.
Made it so you can no longer use a healing cast after healing all broken bones.
Added a new healing cast item that can be used to heal your broken bones in the game.
Made it so animals and zombies can now break your bones depending on the damage received from them.
Added a slash command to reload your character. This is useful for certain situations where your character's stats may be out of wack.