Added a bunch of new features and tweaks. This includes a new combat log menu that let's you keep track of the most recent enemy events in the game.
Fixed a bug where the combat log menu showed incorrect experients counts when leveling up your character.
The combat log menu can now show the experiences gained by all enemy types when killed by your weapons.
The combat log menu can now show the damage stats of all enemy and misc object types when attacked by your weapons.
Made it so the combat log gets cleared upon death or character reload.
Added a keyboard command to open the combat log viewer for looking at the damages received from certain enemy types. You can press shift E to do this.
Made it so animals and zombies are less likely to break your bones if the damage taken by them is lower than 100.
Added some slash commands to reload all of the shields and weapons that are currently present in the game. This is useful for certain situations where your character's shield and or weapon stats may be out of wack.
Made it so the healing cast item heals the last bone that was broken, not just the first one.
Made it so the healing cast item gets recycled after a bone has successfully healed. This is so the same cast can't be used multiple times to continue healing another bone.