Major game enhancements and bug fixes.
Improved the reverb builder, and made it so that you can play a surface from anywhere in the surface list. The checkbox to enable and disable the reverb tester will no longer be present moving forward.
Updated the todo list file with the newest features and bug fixes that should be added and or fixed in future versions of the game.
You can now build surveillance cameras in the builder menu. They will also be written to the map you're currently working on.