Lots of enhancements and bug fixes.
Fixed some issues with the security camera. It will no longer spot you if you aren't in their seeing range, or if it isn't directly facing you when it turns.
Fixed a bunch of issues pertaining to the spire falsely detecting the presents of air when there wasn't one to begin with.