Major game enhancements.
Made it so the loop sounds for elevators will now pause and resume playing upon going in and out of the floors menu.
Made it so all of the door and elevator closing sounds play in 2d instead of playing instationary.
Made it so you can no longer sit and stand, fire weapons, or use quick items while you're inside of a door or an elevator.
Made it so if you are already on the current floor, the elevator won't let you move to the same floor that you chose from the floors menu.
Elevators can now be seen with the spire.
You can now build elevators and elevator floors in the builder menu. They will also be written to the map you're currently working on.
Fixed some password related issues with lockable doors and passages.
Made it so sign sounds don't play while jumping on a map.
Fixed lots of critical issues with bikes.