Major enhancements and bug fixes.
Fixed a bunch of punctuation related issues with speaking the current coordinates of your character, and various other entities. This is also fixed in the object info menu as well.
Fixed some bugs with the auto walking toggle when pressing the letter X.
Made it so your overall jump time is reduced by the weight of your weapons and or shields combined.