Added a bunch of new enhancements and bug fixes.
Added command stacking support to script keys.
Added a slash command to quicly draw weapons of any type.
Added a slash command to reload all of your script keys. This is useful for when you add extra commands that can be triggered with the number row, and you want them to show up in the game.